use model_path instead of model_id for gcs in the cli
The tunix cli does not use the model_id correctly. It provides a gcs/cns path as a model-id, which might not be truly matching a model id especially with custom paths for internal cases. This cl introduces a proper `model_path` field and uses model-id as its true purpose of the unique identifier of the model

PiperOrigin-RevId: 845577928